Add files via upload
This is cheat table is for Rundown 2 for GTFO (Early Access V0.18)
Current features offered:
- Health
- Player Position
Only x/z, meaning only in the directions of North, South, West, East.
- Primary/Secondary Mag Ammo
Do NOT freeze your mag's ammo, else you will not be able to use your melee.
- Equipment Ammo Inventory
- Melee Stun Speed/Weapon Spread
For the last one, it's either it's on or off. There is no customization to weapon spread nor your Stun (hit) Speed.